+++ unit/test/Add-VSTeamIteration.Tests.ps1
@@ -1,0 +1,76 @@
+Set-StrictMode -Version Latest
+
+#region include
+Import-Module SHiPS
+
+$here = Split-Path -Parent $MyInvocation.MyCommand.Path
+$sut = (Split-Path -Leaf $MyInvocation.MyCommand.Path).Replace(".Tests.", ".")
+
+. "$here/../../Source/Classes/VSTeamLeaf.ps1"
+. "$here/../../Source/Classes/VSTeamVersions.ps1"
+. "$here/../../Source/Classes/VSTeamProjectCache.ps1"
+. "$here/../../Source/Classes/VSTeamClassificationNode.ps1"
+. "$here/../../Source/Private/common.ps1"
+. "$here/../../Source/Public/Add-VSTeamClassificationNode"
+. "$here/../../Source/Classes/ProjectCompleter.ps1"
+. "$here/../../Source/Classes/ProjectValidateAttribute.ps1"
+. "$here/../../Source/Public/$sut"
+#endregion
+
+Describe 'Add-VSTeamIteration' {
+
+   $classificationNodeResult = Get-Content "$PSScriptRoot\sampleFiles\classificationNodeResult.json" -Raw | ConvertFrom-Json
+
+   Mock _getInstance { return 'https://dev.azure.com/test' }
+   Mock _getApiVersion { return '5.0-unitTests' } -ParameterFilter { $Service -eq 'Core' }
+
+   Context 'simplest call' {
+      Mock Invoke-RestMethod {
+         #Write-Host $args
+         return $classificationNodeResult 
+      }      
+
+      It 'iteration should return Nodes' {
+         ## Act
+         Add-VSTeamIteration -ProjectName "Public Demo" -Name "MyClassificationNodeName"
+         ## Assert
+         Assert-MockCalled Invoke-RestMethod -Exactly -Times 1 -Scope It -ParameterFilter {
+            $Uri -like "https://dev.azure.com/test/Public Demo/_apis/wit/classificationnodes/iterations*" -and
+            $Uri -like "*api-version=$(_getApiVersion Core)*" -and
+            $Body -like '*"name":*"MyClassificationNodeName"*'
+         }
+      }
+
+      It 'with Path "<Path>" should return Nodes' -TestCases @(
+         @{ Path = 'SubPath' }
+         @{ Path = 'Path/SubPath' }
+     ) {
+         param ($Path)      
+         ## Act
+         Add-VSTeamIteration -ProjectName "Public Demo" -Name "MyClassificationNodeName" -Path $Path
+
+         ## Assert
+         Assert-MockCalled Invoke-RestMethod -Exactly -Times 1 -Scope It -ParameterFilter {
+            $Uri -like "https://dev.azure.com/test/Public Demo/_apis/wit/classificationnodes/iterations/$Path*" -and
+            $Uri -like "*api-version=$(_getApiVersion Core)*" -and
+            $Body -like '*"name":*"MyClassificationNodeName"*'
+         }
+      }
+      
+      It 'with empty Path "<Path>" should return Nodes' -TestCases @(
+         @{Path = "" }
+         @{Path = $null }
+     ) {
+         param ($Path)      
+         ## Act
+         Add-VSTeamIteration -ProjectName "Public Demo" -Name "MyClassificationNodeName" -Path $Path
+
+         ## Assert
+         Assert-MockCalled Invoke-RestMethod -Exactly -Times 1 -Scope It -ParameterFilter {
+            $Uri -like "https://dev.azure.com/test/Public Demo/_apis/wit/classificationnodes/iterations?*" -and
+            $Uri -like "*api-version=$(_getApiVersion Core)*" -and
+            $Body -like '*"name":*"MyClassificationNodeName"*'
+         }
+      }
+   }
+}